Extract and document highlight()
